tests/submission/Baidu/editing/user-interaction-editing-spellcheck.html

<!DOCTYPE html>
<html>
 <head>
  <title>Editing: spellcheck attribute test</title>
  <link rel="author" title="Baidu" href="mailto: guopengcheng@baidu.com"/>
  <link rel="help" href="http://www.w3.org/TR/html5/editing-apis.html#spelling-and-grammar-checking"/>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <div id="log"></div>
 </head>
 <body>
  <textarea id="testText1" spellcheck="true">Test textarea with spellcheck is true</textarea>
  <textarea id="testText2" spellcheck="false">Test textarea with spellcheck is false</textarea>
  <script type="text/javascript">
   test(function() {
    assert_true(document.getElementById("testText1").spellcheck, "check for testText1 spellcheck value");
    assert_false(document.getElementById("testText2").spellcheck, "check for testText2 spellcheck value");
   });
   var testElement = document.createElement("testElement");
   testElement.contentEditable="true";
   testElement.setAttribute("spellcheck", "");
   test(function() {
    assert_true(testElement.spellcheck, "check for testElement.spellcheck value");
   });
   testElement.spellcheck="true";
   test(function() {
    assert_true(testElement.spellcheck, "check for testElement.spellcheck value");
   });
   <!-- SPEC: The spellcheck IDL attribute, on getting, must return true if the element's spellcheck content attribute is in the true state, or if the element's spellcheck content attribute is in the default state and the element's default behavior is true-by-default, or if the element's spellcheck content attribute is in the default state and the element's default behavior is inherit-by-default and the element's parent element's spellcheck IDL attribute would return true; otherwise, if none of those conditions applies, then the attribute must instead return false. -->
   testElement.spellcheck="false";
   test(function() {
    assert_false(testElement.spellcheck, "check for testText2 spellcheck value");
   });
  </script>
 </body>
</html>
